- move sort methods into sidebar component - add header to siedaber - remove sidebar toggle
